Add files via upload
This project focuses on developing an AI model for colorizing grayscale images and videos 
specifically designed for CCTV security systems. The objective is to reduce the storage 
requirements of CCTV systems by promoting the use of grayscale cameras while 
maintaining the accuracy and quality of the footage through AI-based colorization.